Move more code around and add unittests for object count calculation from unloaded tenant
