+++ adapters/repos/db/shard.go
@@ -507,14 +507,6 @@ func shardPathLSM(indexPath, shardName string) string {
 	return path.Join(indexPath, shardName, "lsm")
 }
 
-func shardPathObjectsLSM(indexPath, shardName string) string {
-	return path.Join(shardPathLSM(indexPath, shardName), helpers.ObjectsBucketLSM)
-}
-
-func shardPathDimensionsLSM(indexPath, shardName string) string {
-	return path.Join(shardPathLSM(indexPath, shardName), helpers.DimensionsBucketLSM)
-}
-
 func bucketKeyPropertyLength(length int) ([]byte, error) {
 	return entinverted.LexicographicallySortableInt64(int64(length))
 }
