+++ adapters/repos/db/usage/usage.go
@@ -1,0 +1,222 @@
+package shardusage
+
+import (
+	"context"
+	"encoding/binary"
+	"fmt"
+	"os"
+	"path"
+	"path/filepath"
+	"strings"
+
+	"github.com/sirupsen/logrus"
+	"github.com/weaviate/sroar"
+	"github.com/weaviate/weaviate/adapters/repos/db/helpers"
+	"github.com/weaviate/weaviate/adapters/repos/db/lsmkv"
+	"github.com/weaviate/weaviate/cluster/usage/types"
+	"github.com/weaviate/weaviate/entities/cyclemanager"
+	schemaConfig "github.com/weaviate/weaviate/entities/schema/config"
+)
+
+func shardPathLSM(indexPath, shardName string) string {
+	return path.Join(indexPath, shardName, "lsm")
+}
+
+func shardPathObjectsLSM(indexPath, shardName string) string {
+	return path.Join(shardPathLSM(indexPath, shardName), helpers.ObjectsBucketLSM)
+}
+
+func shardPathDimensionsLSM(indexPath, shardName string) string {
+	return path.Join(shardPathLSM(indexPath, shardName), helpers.DimensionsBucketLSM)
+}
+
+// CalculateUnloadedDimensionsUsage calculates dimensions and object count for an unloaded shard without loading it into memory
+func CalculateUnloadedDimensionsUsage(ctx context.Context, logger logrus.FieldLogger, path, tenantName, targetVector string) (types.Dimensionality, error) {
+	bucketPath := shardPathDimensionsLSM(path, tenantName)
+	strategy, err := lsmkv.DetermineUnloadedBucketStrategyAmong(bucketPath, lsmkv.DimensionsBucketPrioritizedStrategies)
+	if err != nil {
+		return types.Dimensionality{}, fmt.Errorf("determine dimensions bucket strategy: %w", err)
+	}
+
+	bucket, err := lsmkv.NewBucketCreator().NewBucket(ctx,
+		bucketPath,
+		path,
+		logger,
+		nil,
+		cyclemanager.NewCallbackGroupNoop(),
+		cyclemanager.NewCallbackGroupNoop(),
+		lsmkv.WithStrategy(strategy),
+	)
+	if err != nil {
+		return types.Dimensionality{}, err
+	}
+	defer bucket.Shutdown(ctx)
+
+	return CalcTargetVectorDimensionsFromBucket(ctx, bucket, targetVector)
+}
+
+// CalculateUnloadedVectorsMetrics calculates vector storage size for a cold tenant without loading it into memory
+func CalculateUnloadedVectorsMetrics(ctx context.Context, logger logrus.FieldLogger, path, tenantName string, vectorConfig map[string]schemaConfig.VectorIndexConfig) (int64, error) {
+	totalSize := int64(0)
+
+	// For each target vector, calculate storage size using dimensions bucket and config-based compression
+	for targetVector, config := range vectorConfig {
+		err := func() error {
+			bucketPath := shardPathDimensionsLSM(path, tenantName)
+			strategy, err := lsmkv.DetermineUnloadedBucketStrategyAmong(bucketPath, lsmkv.DimensionsBucketPrioritizedStrategies)
+			if err != nil {
+				return fmt.Errorf("determine dimensions bucket strategy: %w", err)
+			}
+
+			// Get dimensions and object count from the dimensions bucket
+			bucket, err := lsmkv.NewBucketCreator().NewBucket(ctx,
+				bucketPath,
+				path,
+				logger,
+				nil,
+				cyclemanager.NewCallbackGroupNoop(),
+				cyclemanager.NewCallbackGroupNoop(),
+				lsmkv.WithStrategy(strategy),
+			)
+			if err != nil {
+				return err
+			}
+			defer bucket.Shutdown(ctx)
+
+			dimensionality, err := CalcTargetVectorDimensionsFromBucket(ctx, bucket, targetVector)
+			if err != nil {
+				return err
+			}
+
+			if dimensionality.Count != 0 && dimensionality.Dimensions != 0 {
+				// Calculate uncompressed size (float32 = 4 bytes per dimension)
+				uncompressedSize := int64(dimensionality.Count) * int64(dimensionality.Dimensions) * 4
+
+				// For inactive tenants, use vector index config for dimension tracking
+				// This is similar to the original shard dimension tracking approach
+				totalSize += int64(float64(uncompressedSize) / helpers.CompressionRatioFromConfig(config, dimensionality.Dimensions))
+			}
+			return nil
+		}()
+		if err != nil {
+			return 0, err
+		}
+	}
+
+	return totalSize, nil
+}
+
+// CalculateUnloadedObjectsMetrics calculates both object count and storage size for a cold tenant without loading it into memory
+func CalculateUnloadedObjectsMetrics(logger logrus.FieldLogger, path, tenantName string) (types.ObjectUsage, error) {
+	// Parse all .cna files in the object store and sum them up
+	totalObjectCount := int64(0)
+	totalDiskSize := int64(0)
+
+	// Use a single walk to avoid multiple filepath.Walk calls and reduce file descriptors
+	if err := filepath.Walk(shardPathObjectsLSM(path, tenantName), func(path string, info os.FileInfo, err error) error {
+		if err != nil {
+			return err
+		}
+
+		// Only count files, not directories
+		if !info.IsDir() {
+			totalDiskSize += info.Size()
+
+			// Look for .cna files (net count additions)
+			if strings.HasSuffix(info.Name(), lsmkv.CountNetAdditionsFileSuffix) {
+				count, err := lsmkv.ReadCountNetAdditionsFile(path)
+				if err != nil {
+					logger.WithField("path", path).WithError(err).Warn("failed to read .cna file")
+					return err
+				}
+				totalObjectCount += count
+			}
+
+			// Look for .metadata files (bloom filters + count net additions)
+			if strings.HasSuffix(info.Name(), lsmkv.MetadataFileSuffix) {
+				count, err := lsmkv.ReadObjectCountFromMetadataFile(path)
+				if err != nil {
+					logger.WithField("path", path).WithError(err).Warn("failed to read .metadata file")
+					return err
+				}
+				totalObjectCount += count
+			}
+		}
+
+		return nil
+	}); err != nil {
+		return types.ObjectUsage{}, err
+	}
+
+	// If we can't determine object count, return the disk size as fallback
+	return types.ObjectUsage{
+		Count:        totalObjectCount,
+		StorageBytes: totalDiskSize,
+	}, nil
+}
+
+// calcTargetVectorDimensionsFromBucket calculates dimensions and object count for a target vector from an LSMKV bucket
+func CalcTargetVectorDimensionsFromBucket(ctx context.Context, b *lsmkv.Bucket, targetVector string,
+) (types.Dimensionality, error) {
+	dimensionality := types.Dimensionality{}
+
+	if err := lsmkv.CheckExpectedStrategy(b.Strategy(), lsmkv.StrategyMapCollection, lsmkv.StrategyRoaringSet); err != nil {
+		return dimensionality, fmt.Errorf("calcTargetVectorDimensionsFromBucket: %w", err)
+	}
+
+	nameLen := len(targetVector)
+	expectedKeyLen := nameLen + 4 // vector name + uint32
+	var k []byte
+
+	switch b.Strategy() {
+	case lsmkv.StrategyMapCollection:
+		// Since weaviate 1.34 default dimension bucket strategy is StrategyRoaringSet.
+		// For backward compatibility StrategyMapCollection is still supported.
+
+		c := b.MapCursor()
+		defer c.Close()
+
+		var v []lsmkv.MapPair
+		if nameLen == 0 {
+			k, v = c.First(ctx)
+		} else {
+			k, v = c.Seek(ctx, []byte(targetVector))
+		}
+		for ; k != nil; k, v = c.Next(ctx) {
+			// for named vectors we have to additionally check if the key is prefixed with the vector name
+			if len(k) != expectedKeyLen || !strings.HasPrefix(string(k), targetVector) {
+				break
+			}
+
+			dimLength := binary.LittleEndian.Uint32(k[nameLen:])
+			if dimLength > 0 && (dimensionality.Dimensions == 0 || dimensionality.Count == 0) {
+				dimensionality.Dimensions = int(dimLength)
+				dimensionality.Count = len(v)
+			}
+		}
+	default:
+		c := b.CursorRoaringSet()
+		defer c.Close()
+
+		var v *sroar.Bitmap
+		if nameLen == 0 {
+			k, v = c.First()
+		} else {
+			k, v = c.Seek([]byte(targetVector))
+		}
+		for ; k != nil; k, v = c.Next() {
+			// for named vectors we have to additionally check if the key is prefixed with the vector name
+			if len(k) != expectedKeyLen || !strings.HasPrefix(string(k), targetVector) {
+				break
+			}
+
+			dimLength := binary.LittleEndian.Uint32(k[nameLen:])
+			if dimLength > 0 && (dimensionality.Dimensions == 0 || dimensionality.Count == 0) {
+				dimensionality.Dimensions = int(dimLength)
+				dimensionality.Count = v.GetCardinality()
+			}
+		}
+	}
+
+	return dimensionality, nil
+}
